add placeholders for ui animations, document canvas components
