+++ src/components/modal/Modal.tsx
@@ -50,21 +50,43 @@ const Modal: React.FC<ModalProps> = ({
 	return (
 		<div
 			aria-hidden={!isOpen}
-			className="overflow-y-auto backdrop-blur-sm fixed top-0 right-0 left-0 z-50 flex justify-center items-center w-full h-full max-h-full bg-black bg-opacity-50"
-		> 
+			className="overflow-auto backdrop-blur-sm fixed top-0 right-0 left-0 z-50 flex justify-center items-center w-full h-full max-h-full bg-black bg-opacity-50"
+		>
+			{/* Add animation:
+				Overlay fade-in/fade-out or blur transition
+			*/}
+
 			<div
 			ref={modalRef}
 			className="relative flex flex-col px-2 w-5/6 md:w-1/2 lg:w-1/3 max-w-lg sm:max-w-full max-h-full overflow-auto rounded-lg shadow bg-background"
 			>
+				{/* Add animation:
+					Modal box entrance (e.g., slide-in, scale-up, fade-in)
+				*/}
+
 				{/* Modal Header */}
 				<div className="md:flex text-center md:text-left items-center justify-between pt-3 md:p-3">
 					<Header text={title} textSize="text-2xl" hasGradient gradientColors={headerColors}/>
+					{/* Close button */}
 					{hasCloseButton ? (<CloseModalButton onClick={toggleModal}/>) : (<></>)}
+					{/* Add animation:
+						Some kind of hover effect on the close button.
+						Modify CloseModalButton component to do this.
+						Suggestions: Scale up on hover?
+					*/}
 				</div>
 				{/* Modal Content */}
 				<div className='flex-grow px-3 pt-3 md:pt-0 pb-3 space-y-4 text-text'>
+					{/* Add animation:
+						Modal content entrance and exit (e.g., fade-in)
+						Optional, do you need animation for the content if the box is animated?
+					*/
+					}
 					{children}
 				</div>
+				{/* Add animation:
+					Modal box exit (e.g., slide-out, scale-down, fade-out)
+				*/}	
 
 			</div>
 		</div>
